parse_base/requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
lxml==4.7.1
python-dotenv==0.19.2
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.7
psycopg2-binary==2.9.3
SQLAlchemy==1.4.35
aiofiles==0.8.0
aiohttp==3.8.1
